使用Python构建RAG系统/rag/requirements.txt
sentence-transformers>=3.3.0
chromadb>=0.5.0
python-dotenv>=1.0.1
google-genai>=0.3.0
torch>=2.2.0,<=2.8.0
torchvision
torchaudio
openai>=1.40.0
jiter>=0.4.0
overrides>=7.0.0
numpy<2.0.0 